moved "timers" into their own file
